requirements.txt
asttokens==3.0.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
easydict==1.13
einops==0.8.0
et_xmlfile==2.0.0
executing==2.2.1
filelock==3.16.0
fonttools==4.60.1
fsspec==2024.9.0
gdown==5.2.0
gmpy2==2.2.1
h2==4.1.0
h5py==3.15.1
hpack==4.0.0
huggingface-hub==0.24.7
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflate64==1.0.3
ipdb==0.13.13
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.2.1
mpmath==1.3.0
multivolumefile==0.2.3
networkx==3.3
numpy==2.1.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==24.1
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
POT==0.9.6.post1
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py7zr==1.0.0
pybcj==1.0.6
pycparser==2.23
pycryptodomex==3.23.0
Pygments==2.19.2
pyparsing==3.2.5
pyppmd==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzstd==0.17.0
requests==2.32.5
safetensors==0.4.5
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.14.1
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.2
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.10.4
timm==1.0.9
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.15.0
tzdata==2025.2
umap==0.1.1
urllib3==2.5.0
wcwidth==0.2.14
zstandard==0.23.0
